move EngineBufferScaler classes to their own folder
