+++ src/engine/positionscratchcontroller.cpp
@@ -1,7 +1,7 @@
 #include <QtDebug>
 
 #include "engine/positionscratchcontroller.h"
-#include "engine/enginebufferscale.h" // for MIN_SEEK_SPEED
+#include "engine/bufferscalers/enginebufferscale.h" // for MIN_SEEK_SPEED
 #include "util/math.h"
 
 class VelocityController {
